+++ cli/lsp/config.rs
@@ -8,7 +8,7 @@ use crate::args::LintOptions;
 use crate::args::DENO_FUTURE;
 use crate::cache::FastInsecureHasher;
 use crate::file_fetcher::FileFetcher;
-use crate::lockfile::CliLockfile;
+use crate::args::CliLockfile;
 use crate::lsp::logging::lsp_warn;
 use crate::tools::lint::get_configured_rules;
 use crate::tools::lint::ConfiguredRules;
