+++ cli/mainrt.rs
@@ -15,7 +15,6 @@ mod errors;
 mod file_fetcher;
 mod http_util;
 mod js;
-mod lockfile;
 mod node;
 mod npm;
 mod resolver;
